Split growing API test file
