+++ tests/__init__.py
@@ -1,1 +1,3 @@
 from .test_version import *
+from .test_api_browse import *
+from .test_api_mod import *
