+++ tests/test_api_mod.py
@@ -98,27 +98,3 @@ def test_api_mod(client: 'FlaskClient[Response]') -> None:
     assert user_resp.json['description'] == 'Test author of a test mod', 'Description should match'
     assert user_resp.json['forumUsername'] == 'TestForumUser', 'Forum name should match'
     assert user_resp.json['mods'][0]['name'] == 'Test Mod', 'Mod should be returned'
-
-
-@pytest.mark.usefixtures("client")
-def test_api_browse(client: 'FlaskClient[Response]') -> None:
-    # Arrange is handled by the fixture
-
-    # Act
-    browse_resp = client.get('/api/browse')
-    new_resp = client.get('/api/browse/new')
-    top_resp = client.get('/api/browse/top')
-    featured_resp = client.get('/api/browse/featured')
-
-    # Assert
-    assert browse_resp.status_code == status.HTTP_200_OK, 'Request should succeed'
-    assert browse_resp.data == b'{"total":0,"count":30,"pages":1,"page":1,"result":[]}', 'Should be a simple empty db'
-
-    assert new_resp.status_code == status.HTTP_200_OK, 'Request should succeed'
-    assert new_resp.data == b'[]', 'Should return empty list'
-
-    assert top_resp.status_code == status.HTTP_200_OK, 'Request should succeed'
-    assert top_resp.data == b'[]', 'Should return empty list'
-
-    assert featured_resp.status_code == status.HTTP_200_OK, 'Request should succeed'
-    assert featured_resp.data == b'[]', 'Should return empty list'
